8341238: G1: Refactor G1Policy to move collection set selection methods into G1CollectionSet
