+++ src/hotspot/share/gc/g1/g1Policy.hpp
@@ -335,27 +335,7 @@ public:
 
   // Amount of allowed waste in bytes in the collection set.
   size_t allowed_waste_in_collection_set() const;
-  // Calculate and fill in the initial, optional and pinned old gen candidate regions from
-  // the given candidate list and the remaining time.
-  // Returns the remaining time.
-  double select_candidates_from_marking(G1CollectionCandidateList* marking_list,
-                                        double time_remaining_ms,
-                                        G1CollectionCandidateRegionList* initial_old_regions,
-                                        G1CollectionCandidateRegionList* optional_old_regions,
-                                        G1CollectionCandidateRegionList* pinned_old_regions);
 
-  void select_candidates_from_retained(G1CollectionCandidateList* retained_list,
-                                       double time_remaining_ms,
-                                       G1CollectionCandidateRegionList* initial_old_regions,
-                                       G1CollectionCandidateRegionList* optional_old_regions,
-                                       G1CollectionCandidateRegionList* pinned_old_regions);
-
-  // Calculate the number of optional regions from the given collection set candidates,
-  // the remaining time and the maximum number of these regions and return the number
-  // of actually selected regions in num_optional_regions.
-  void calculate_optional_collection_set_regions(G1CollectionCandidateRegionList* optional_old_regions,
-                                                 double time_remaining_ms,
-                                                 G1CollectionCandidateRegionList* selected);
 
 private:
 
@@ -423,12 +403,12 @@ private:
 
   size_t desired_survivor_size(uint max_regions) const;
 
+public:
   // Fraction used when predicting how many optional regions to include in
   // the CSet. This fraction of the available time is used for optional regions,
   // the rest is used to add old regions to the normal CSet.
   double optional_prediction_fraction() const { return 0.2; }
 
-public:
   // Fraction used when evacuating the optional regions. This fraction of the
   // remaining time is used to choose what regions to include in the evacuation.
   double optional_evacuation_fraction() const { return 0.75; }
